# Conflicts:
#	ace/ace-jxb/ace-jxb-web/src/main/webapp/content/service/counselorLevel/js/counselorLevel.js
#	ace/ace-jxb/ace-jxb-web/src/main/webapp/dynamic/service/counselorLevel/index.jsp

+++ ace/ace-jxb/ace-jxb-web/src/main/webapp/www/view/order/js/act.js
@@ -101,7 +101,7 @@ function changeSex(obj,value){
 function changeType(obj, priceStr, id){
 	$(obj).removeClass("unactive").addClass("active");
 	$(obj).parent().siblings().children().removeClass("active").addClass("unactive");
-    unitPrice = parseInt(priceStr);
+    unitPrice = parseFloat(priceStr);
     commodityId = id;
     var num = parseInt($("#num").text());
     $("#totalMoney").text(num * unitPrice);
